Continue to improve auto-complete interaction
Words in comments are now ignored. Most of my auto-complete needs are from
identifiers so any words from comments are generally just noise.

Also start auto-completing snippets as well. Since an auto-complete selected
might be a snippet automatically activate the snippet. It might be nice at
some future point to indicate what list a auto-complete comes from but that
could just be noise so keeping it simple for now.