Attempt to improve odd opening of files due to this module
It feels like a race condition. I will open a new file but it will think the
old one is the active one still an close the new one instead of the old one.
Adddinig a timeout seems to help but not perfectly. Even a couple seconds later
it is confused on what is the current active view and close the one I opened.

But since better will commit for now. I wonder if I should basically change this
from "ignore active" to "ignore active and the one that triggered the event".