Add logger to help debugging
It is often difficult for me to tell what is happening and I'm not
sure how to use a debugger with this embedded interpreter. Setting
`ui.statusbar_text` can help some but it only hold the last message.
Using `ui.print` can output better into but it sometimes interferes
with the process since it opens a buffer.

This way I can do log messages completely in the background with:

    textadept.log('my_module', 'Some message')

This will write to `~/.textadept/activity.log`. The message is prefixed
with the specified module name in brackets (for easy filter) as well
as a timestamp (to help determine relevant messages).