Hopefully a more reliable version of this module.
Just a timeout wasn't solving it as it often still messed with the buffers
when it should not have. Tracked it down to two issues.

I think the events being triggered are in their own thread meaning they often
overlap causing problems with each other. This is solved by only running one
at a time. If any others requested while one is running then ignore it. Still
have a slight delay to give other modules a chance to "claim" a buffer before
cleanup but it's not strictly necessary.

The second issue is something internal to textadept causes the unrelated buffers
on unrelated views to change. I've pinged on the project but can also work
around it for now but just remembering the association and then ensuring that
is the assocation when done.