Refine closing column and add column swap
Not only was my original close function not working well I don't think it is
what we want. If I want to reduce from 5 to 4 columns I want to be specific
about what file I don't want anymore. Therefore changed the semantics so that
instead of setting the column count it's more toggling the presense of a
specific column.

Examples:

* I have 5 columns I an toggle column 2. I will have 4 columns with the old
  column 2 removed.
* If I toggle the column greater than my column column it will add the columns
  necessary to show the column I'm interested in.

I also found when working that I often want to keep a file but I just want it
in a different place. I could switch to that column, then open the file I'm
interested in. But I decided just hitting `ctrl-shift-#` to swap the content
is the current column with the target column might be better as it then also
does something with the old position that seems reasonable. And of course if
I don't want the file it was swapped with anymore I could follow it up with
toggling the old column.