Improve auto-complete when nested in snippet
Finally figured out that my issue was because I was unconditonally inserting
a snippet on auto-complete. Even if it matched nothing the impl of
snippet:insert made it move to the next placeholder (in hopes it is the first
one) but that ended up moving it to the next placeholder in the parent snippet.

By checking first that I actually chose a snippet and only auto-inserting
the snippet if my auto-complete is a snippet it works better and now my control
over the keys while the auto-commplete shows is no longer needed.