Some reorganization
I don't like that all files were named `init.lua`. Found there is a search
pattern for where the module is just a single file in the `modules/` directory
so moving to that unless a module really needs a full directory.

Also moved some of the stuff in `init.lua` into their own new modules to try
to organize a bit better. Might do more stuff in the future but it's a good
start.