Disable sessions
I don't find them that useful. Generally if I'm still deep in a task
then I'll just leave the window open and my computer on. If I've closed
my window it's likely because I've moved onto a different task.

Even if I did shut things down and want to get back where I was I
find the work of re-opening less than dealing with cruft that needs
to be constantly closed down when there is a task change.