Resolve single item problem
I solved this before but started getting it again. In b6ecadd I was
able to remove a lot of complication by handling snippet auto-completion
better so I removed that complication.

The problem was I removed too much. I still need the code to cancel
the auto-complete if it is shown as I noted in this discussion:

https://github.com/orbitalquark/textadept/discussions/283#discussioncomment-3928895

Only now since I'm not using the stock auto-complete but instead my
multi-autocomplete I can really solve it in the more correct place rather
than just in a hook that runs before the correct place.

In addition to moving that cancel into the multi-autocomplete I also
removed the guard for the scenario where no lists are passed in. It seems
more correct that the function would handle this data rather than the
caller having to guard against sending bad data.