Use different key combo for quick open
Ctrl-O is fine for Mac since Cmd-O does the file open but for PC it will replace
the file open and I think we still want that for when we open files outside a
project.

Looked at Atom to see it's keyboard shortcut. Although already used there is
a better alternative so we can steal it. Now we have both OS open and quick
open for both operating systems.