Fix quick open duplicate checking
My logic was wrong. I was doing a lookup as if it were a hash where the key
is the root but it's an array where the key is an index. Use my contains
utility method instead.

Decided to simplify that utility method for the common case. If the evaluator
is not a function then auto-wrap it in a simple equality comparison function.
This simplified several other modules.