Allow easily copying of current path to clipboard
Most people tend to have their editor run some command against a file. But that
means for each command you want you need to setup something in your editor
and you still don't get a full shell.

I already have a shell which is excellent for running commands. I just often
need to copy/paste the file path so the command knows what file to operate on.
This gives me that.

Either shortcut will work just fine for most purposes. The longer one will
always work but it can also produce verbose paths as it is the full path from
the root. The shorter one assumes your shell is in the root of your project
so we can clip off the project portion of the path.

If the project root cannot be determined it falls back on the full path from
the root.