Simplify no_tab module
The developer pointed to where the problem is at. I also could not think of a
good way to solve it within the application and therefore did not submit an
upstream patch. Instead the feature causing the problem is not useful when also
using this module. Therefore just auto-disable the feature by reaching into the
internals a bit. :)