Move away from git submodules for 3rd party modules
Instead just doing simple shell script. The submodules didn't work well because
it grabbed the source which isn't always pure Lua. Sometimes it has compiled
code that needs to be built.

By grabbing the pre-built package we get the module with the built binaries
which makes the whole process easier.