Remove LSP
I played around with it and overall for Ruby didn't find it that useful. There
also doesn't seem to be a good server for JavaScript which is my other heavy
language.

If I start using another language heavily that could benefit from it then I'm
not opposed to putting it back. But for now it doesn't add enough value.